Rename _layers_list/view
