+++ napari/components/_layers_list/qtlayerlist.py
